add test for adding new expense

+++ app/src/main/java/com/nsnik/nrs/jaron/view/fragments/dialogs/AddExpenseFragment.kt
@@ -79,7 +79,7 @@ class AddExpenseFragment : DialogFragment() {
 
     private fun validateEntries(): Boolean {
         return validateFrom(
-            textViewToString(newExpenseValue).toDouble(),
+            textViewToString(newExpenseValue),
             textViewToString(newExpenseTitle),
             textViewToString(newExpenseDescription),
             Calendar.getInstance().time,
